+++ client/tests/acceptance/happy-path/happy-path.spec.ts
@@ -1,14 +1,17 @@
 import { describe, it, xit, expect } from "../../test-support";
-import tracksFor from "../../test-support/data-storage";
+import tracksFor from "./data-storage";
 import { ITrack } from "../../test-support/types";
 import { join } from 'path';
 
 import {
   pageObject,
+} from "../../test-support/page-object";
+
+import {
   setupPage,
   buildUrl,
   mockData
-} from "../../test-support/page-object";
+} from "../../test-support/test-helpers";
 
 describe("Acceptance | Happy path |", () => {
   xit("is a basic test with the page", async ({ page }) => {
