+++ client/tests/package.json
@@ -1,24 +1,15 @@
 {
   "name": "ngb-tests",
   "version": "1.0.0",
-  "description": "",
   "type": "module",
   "private": true,
-  "scripts": {
-    "test": "npx folio ./test --param browserName=chromium",
-    "test:visual": "cross-env HEADFUL=true yarn test",
-    "start-server": "yarn --cwd ./../client serve:test",
-    "start-test": "yarn --cwd . test",
-    "ci": "start-server-and-test start-server http-get://localhost:8080 start-test"
-  },
-  "author": "",
-  "dependencies": {
+  "peerDependencies": {
     "@playwright/test": "^0.180.0",
     "cross-env": "^7.0.3",
     "folio": "^0.3.17",
     "start-server-and-test": "^1.12.0"
   },
   "volta": {
-    "node": "12.20.1"
+    "node": "14.15.5"
   }
 }
