+++ client/tests/test-support/data-storage/index.ts
@@ -1,287 +1,6 @@
-import { ITrack } from "../types";
-
-const tracks: {
-  [key: string]: ITrack[];
-} = {
-  "is a basic test with the page": [
-    { b: "DM6", p: "Fruitfly", f: "REFERENCE", projectIdNumber: 1 },
-    {
-      b: "agnts3.09-28.trim.dm606.realign.bam",
-      p: "Fruitfly",
-      f: "BAM",
-      projectIdNumber: 1,
-    },
-    {
-      b: "agnX1.09-28.trim.dm606.realign.bam",
-      p: "Fruitfly",
-      f: "BAM",
-      projectIdNumber: 1,
-    },
-    {
-      b: "agnX1.09-28.trim.dm606.realign.vcf",
-      p: "Fruitfly",
-      f: "VCF",
-      projectIdNumber: 1,
-    },
-    {
-      b: "CantonS.09-28.trim.dm606.realign.bam",
-      p: "Fruitfly",
-      f: "BAM",
-      projectIdNumber: 1,
-    },
-    {
-      b: "CantonS.09-28.trim.dm606.realign.vcf",
-      p: "Fruitfly",
-      f: "VCF",
-      projectIdNumber: 1,
-    },
-    { b: "DM6_Genes", p: "", l: true, f: "GENE" },
-  ],
-  "it able to test mouse movements": [
-    {
-      b: "agnts3.09-28.trim.dm606.realign.bam",
-      f: "BAM",
-      h: 250,
-      p: "Fruitfly",
-      s: {
-        a: true,
-        aa: true,
-        c: "noColor",
-        c1: true,
-        d: true,
-        g1: "default",
-        i: true,
-        m: true,
-        r: 1,
-        s1: false,
-        s2: true,
-        s3: false,
-        v1: false,
-        cdm: "Bar Graph",
-        cls: false,
-        csm: "default",
-      },
-      projectIdNumber: 1,
-    },
-    {
-      b: "agnX1.09-28.trim.dm606.realign.bam",
-      f: "BAM",
-      h: 250,
-      p: "Fruitfly",
-      s: {
-        a: true,
-        aa: true,
-        c: "noColor",
-        c1: true,
-        d: true,
-        g1: "default",
-        i: true,
-        m: true,
-        r: 1,
-        s1: false,
-        s2: true,
-        s3: false,
-        v1: false,
-        cdm: "Bar Graph",
-        cls: false,
-        csm: "default",
-      },
-      projectIdNumber: 1,
-    },
-    {
-      b: "DM6",
-      f: "REFERENCE",
-      h: 100,
-      p: "Fruitfly",
-      s: { rt: false, rsfs: true, rsrs: false },
-      projectIdNumber: 1,
-    },
-    {
-      b: "agnX1.09-28.trim.dm606.realign.vcf",
-      f: "VCF",
-      h: 70,
-      p: "Fruitfly",
-      s: { v: "Collapsed" },
-      projectIdNumber: 1,
-    },
-    {
-      b: "CantonS.09-28.trim.dm606.realign.bam",
-      f: "BAM",
-      h: 250,
-      p: "Fruitfly",
-      s: {
-        a: true,
-        aa: true,
-        c: "noColor",
-        c1: true,
-        d: true,
-        g1: "default",
-        i: true,
-        m: true,
-        r: 1,
-        s1: false,
-        s2: true,
-        s3: false,
-        v1: false,
-        cdm: "Bar Graph",
-        cls: false,
-        csm: "default",
-      },
-      projectIdNumber: 1,
-    },
-    {
-      b: "CantonS.09-28.trim.dm606.realign.vcf",
-      f: "VCF",
-      h: 70,
-      p: "Fruitfly",
-      s: { v: "Collapsed" },
-      projectIdNumber: 1,
-    },
-    {
-      b: "DM6_Genes",
-      f: "GENE",
-      h: 158,
-      l: true,
-      p: "",
-      s: { g: "collapsed" },
-    },
-  ],
-  "matches barchart graph": [
-    {
-      b: "h37",
-      p: "h37_data",
-      f: "REFERENCE",
-      projectIdNumber: 1,
-      h: 100,
-      s: { rt: false, rsfs: true, rsrs: false },
-    },
-    {
-      b: "sv_sample_2.bw",
-      h: 86,
-      p: "h37_data",
-      projectIdNumber: 1,
-      s: {
-        cdm: "Bar Graph",
-        cls: false,
-        csm: "default",
-        csf: 100,
-        cst: 900,
-        wigColors: { negative: 16731471, positive: 3265094 },
-      },
-      n: "sv_sample_2.bw",
-      f: "WIG",
-    },
-  ],
-  "matches distribution graph": [
-    {
-      b: "h37",
-      p: "h37_data",
-      f: "REFERENCE",
-      projectIdNumber: 1,
-      h: 100,
-      s: { rt: false, rsfs: true, rsrs: false },
-    },
-    {
-      b: "Homo_sapiens.GRCh37.gtf.gz",
-      h: 84,
-      l: true,
-      p: "",
-      s: { g: "expanded" },
-      f: "GENE",
-    },
-    {
-      b: "sv_sample_2.bw",
-      h: 150,
-      p: "h37_data",
-      projectIdNumber: 1,
-      s: {
-        cdm: "Bar Graph",
-        cls: false,
-        csm: "default",
-        csf: 100,
-        cst: 900,
-        wigColors: { negative: 16731471, positive: 3265094 },
-      },
-      n: "sv_sample_2.bw",
-      f: "WIG",
-    },
-  ],
-  "works for track renderer": [
-    {
-      b: "dm6",
-      p: "dm6_data",
-      f: "REFERENCE",
-      projectIdNumber: 2,
-      h: 100,
-      s: { rt: false, rsfs: true, rsrs: false },
-    },
-    {
-      b: "agnX1.09-28.trim.dm606.realign.bam",
-      h: 576,
-      p: "dm6_data",
-      projectIdNumber: 2,
-      s: {
-        a: true,
-        aa: true,
-        c: "insertSize",
-        c1: true,
-        d: true,
-        g1: "default",
-        he: { fontSize: "16px" },
-        i: true,
-        m: true,
-        r: 2,
-        s1: true,
-        s2: true,
-        s3: false,
-        v1: false,
-        cdm: "Bar Graph",
-        cls: false,
-        csm: "default",
-      },
-      n: "agnX1.09-28.trim.dm606.realign.bam",
-      f: "BAM",
-    },
-  ],
-  "matches with disabled downsampling": [
-    {
-      b: "dm6",
-      p: "dm6_data",
-      f: "REFERENCE",
-      projectIdNumber: 2,
-      h: 100,
-      s: { rt: false, rsfs: true, rsrs: false },
-    },
-    {
-      b: "agnX1.09-28.trim.dm606.realign.bam",
-      h: 576,
-      p: "dm6_data",
-      projectIdNumber: 2,
-      s: {
-        a: true,
-        aa: true,
-        c: "insertSize",
-        c1: true,
-        d: true,
-        g1: "default",
-        he: { fontSize: "16px" },
-        i: true,
-        m: true,
-        r: 2,
-        s1: true,
-        s2: true,
-        s3: false,
-        v1: false,
-        cdm: "Bar Graph",
-        cls: false,
-        csm: "default",
-      },
-      n: "agnX1.09-28.trim.dm606.realign.bam",
-      f: "BAM",
-    },
-  ],
-};
-
+import type { ITrack } from "../types";
+const tracks = {};
 export default function tracksFor(key: keyof typeof tracks): ITrack[] {
-  return tracks[key];
+    return tracks[key];
 }
+
